Image, ColorImage, GrayImageクラスを個別ファイルに分けた
