+++ freshman2019/camera/__init__.py
@@ -1,2 +1,4 @@
 from .camera import *
-from .cv import *
+from .image import Image
+from .gray_image import GrayImage
+from .color_image import ColorImage
